refactor: move stubs and factory to mock-context-provider
- Moved StubRecentlyVisitedRangesService to mock-context-provider.ts
- Moved StubRecentlyEditedTracker to mock-context-provider.ts
- Moved createProviderForTesting factory to mock-context-provider.ts
- Simplified ghost-provider-tester.ts imports
- Better organization: all mocks/stubs in one place
- All tests passing (25/25)
